+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   ([rhbz#2151164], [ghissue#604])
 - Allow time values in stonith-watchdog-time property ([rhbz#2158790])
 - Enable/Disable sbd when cluster is not running ([rhbz#2166249])
+- Confusing error message in `pcs constraint ticket add` command
+  ([rhbz#2168617], [ghpull#559])
 
 ### Changed
 - Resource/stonith agent self-validation of instance attributes is now
@@ -22,11 +24,13 @@
   ([rhbz#2159454])
 
 [ghissue#604]: https://github.com/ClusterLabs/pcs/issues/604
+[ghpull#559]: https://github.com/ClusterLabs/pcs/pull/559
 [rhbz#2151164]: https://bugzilla.redhat.com/show_bug.cgi?id=2151164
 [rhbz#2151524]: https://bugzilla.redhat.com/show_bug.cgi?id=2151524
 [rhbz#2158790]: https://bugzilla.redhat.com/show_bug.cgi?id=2158790
 [rhbz#2159454]: https://bugzilla.redhat.com/show_bug.cgi?id=2159454
 [rhbz#2166249]: https://bugzilla.redhat.com/show_bug.cgi?id=2166249
+[rhbz#2168617]: https://bugzilla.redhat.com/show_bug.cgi?id=2168617
 
 
 ## [0.11.4] - 2022-11-21